Fix bugs found by SonarCloud

+++ src/ruuvi_endpoint_5.c
@@ -393,8 +393,8 @@ bool re_5_check_format (const uint8_t * const p_buffer)
         return false;
     }
 
-    const uint16_t manufacturer_id = ( (uint16_t)
-                                       p_buffer[RE_5_RAW_PACKET_MANUFACTURER_ID_OFFSET_HI] << RE_5_BYTE_1_SHIFT)
+    const uint16_t manufacturer_id = (uint16_t) ( (uint16_t)
+                                     p_buffer[RE_5_RAW_PACKET_MANUFACTURER_ID_OFFSET_HI] << RE_5_BYTE_1_SHIFT)
                                      + p_buffer[RE_5_RAW_PACKET_MANUFACTURER_ID_OFFSET_LO];
 
     if (RE_5_RAW_PACKET_MANUFACTURER_ID_VAL != manufacturer_id)
